requirements.txt
autopep8==1.4.3
dj-database-url==0.5.0
Django==2.1.5
django-taggit==0.23.0
djangorestframework==3.9.1
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pytz==2018.9
